Fix event emitter for fs when writing outside of project
File is null when outside of the project directory. The emitter was
causing the write creation to fire (correctly) but the event handler
in fsExtensions was trying to write data to the explorer when no
file existed there (in manaco). Exposed fsEvents in case applications
want to also subscribe to events.