Allow filesystem to automatically coerce async to sync
When trying to mount the filesystem to emscripten I found that
the emscripten FS API is entirely synchronous for node but doesn't
call the "Sync" methods. When this occurs we get a type error that
callback is missing for the async call. We can check on the FS
if the equivelant sync call is available and then call it
automatically so that the async API on FS also works sync when
you do supply a callback.